// Front/src/app/actividade/model/actividade-new.ts
export class Actividade {
  id: number; 
  nome: string;
  descricao: string;
  user_id: number;
  estado: number;
  data_inicio: any;
  data_fim: any;
  duracao: number;
  responsavel: number;
  actividade_tipo_id:number
  despacho: string;
  gd_documento_id: number;

}
//Dados que encontrei
// nome: string;
// cor: string;
// estado: number;
// descricao: string; 
// data_inicio: string;
// data_fim: string;
// duracao:number
// actividade_tipo_id:number
// responsavel:number
// area_despacho:any
// user_despacho:any